requirements.txt
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
nltk>=3.8.0
pyyaml>=6.0
